(WIP) Implement storage key/type retrieval
Very important that we record type information in the path of the file.
This makes it super easy to serialize/deserialize all keys at will
without any extra information stored. Being able to do this with a
single scan cuts down on the risk of tracking files and eventually
losing one.