Fix directory parsing & pass back type array size
Apparently strtok modifies thet string and only supports one sequence at
a time. Therefore, we use strtok_r with a context pointer such that it
can differentiate between the two parsing instances.

Caller of storage_getAllKeys needs to know the size of the type array so
a size pointer is included.